+++ app/src/main/java/woowacourse/shopping/data/source/ProductDataSource.kt
@@ -5,7 +5,7 @@ import woowacourse.shopping.data.model.ProductData
 interface ProductDataSource {
     fun findByPaged(page: Int): List<ProductData>
 
-    fun findById(id: Int): ProductData
+    fun findById(id: Long): ProductData
 
     fun isFinalPage(page: Int): Boolean
 }
