+++ app/src/main/java/woowacourse/shopping/data/source/ShoppingCartProductIdDataSource.kt
@@ -3,7 +3,7 @@ package woowacourse.shopping.data.source
 import woowacourse.shopping.data.model.ProductIdsCountData
 
 interface ShoppingCartProductIdDataSource {
-    fun findByProductId(productId: Int): ProductIdsCountData?
+    fun findByProductId(productId: Long): ProductIdsCountData?
 
     fun loadPaged(page: Int): List<ProductIdsCountData>
 
@@ -11,13 +11,13 @@ interface ShoppingCartProductIdDataSource {
 
     fun isFinalPage(page: Int): Boolean
 
-    fun addedNewProductsId(productIdsCountData: ProductIdsCountData): Int
+    fun addedNewProductsId(productIdsCountData: ProductIdsCountData): Long
 
-    fun removedProductsId(productId: Int): Int
+    fun removedProductsId(productId: Long): Long
 
-    fun plusProductsIdCount(productId: Int)
+    fun plusProductsIdCount(productId: Long)
 
-    fun minusProductsIdCount(productId: Int)
+    fun minusProductsIdCount(productId: Long)
 
     fun clearAll()
 }
