+++ app/src/main/java/woowacourse/shopping/domain/repository/ProductIdsCountRepository.kt
@@ -3,17 +3,17 @@ package woowacourse.shopping.domain.repository
 import woowacourse.shopping.domain.model.ProductIdsCount
 
 interface ProductIdsCountRepository {
-    fun findByProductId(productId: Int): ProductIdsCount
+    fun findByProductId(productId: Long): ProductIdsCount
 
     fun loadAllProductIdsCounts(): List<ProductIdsCount>
 
-    fun addedProductsId(productIdsCount: ProductIdsCount): Int
+    fun addedProductsId(productIdsCount: ProductIdsCount): Long
 
-    fun removedProductsId(productId: Int): Int
+    fun removedProductsId(productId: Long): Long
 
-    fun plusProductsIdCount(productId: Int)
+    fun plusProductsIdCount(productId: Long)
 
-    fun minusProductsIdCount(productId: Int)
+    fun minusProductsIdCount(productId: Long)
 
     fun clearAll()
 }
