+++ app/src/test/java/woowacourse/shopping/source/FakeProductDataSource.kt
@@ -10,7 +10,7 @@ class FakeProductDataSource(
 ) : ProductDataSource {
     override fun findByPaged(page: Int): List<ProductData> = pagingStrategy.loadPagedData(page, allProducts)
 
-    override fun findById(id: Int): ProductData =
+    override fun findById(id: Long): ProductData =
         allProducts.find { it.id == id }
             ?: throw NoSuchElementException("there is no product with id: $id")
